Move package under haystack_integrations
